refactor: decompose monolithic routes.ts into modular routes
- Extracted ML bulk endpoint to ml.routes.ts
- Extracted export endpoint to exports.routes.ts
- Extracted analytics endpoint to analytics.routes.ts
- Removed duplicated /api/assessments route blocks

Closes #660
