+++ server/routes.ts
@@ -1,3 +1,6 @@
+import mlRouter from "./routes/ml.routes";
+import exportsRouter from "./routes/exports.routes";
+import analyticsRouter from "./routes/analytics.routes";
 import type { Express } from "express";
 import type { Server } from "http";
 import authRouter from "./routes/auth.routes";
@@ -409,1232 +412,9 @@ export async function registerRoutes(
     res.json({ token });
   });
 
-  app.post(
-    api.assessments.preview.path,
-    requireAuth,
-    requireVerified,
-    previewLimiter,
-    async (req, res) => {
-      const input = api.assessments.preview.input.parse(req.body);
-      const tempFile = path.join(
-        os.tmpdir(),
-        `${randomUUID()}.json`
-      );
-
-      try {
-        await writeFile(tempFile, JSON.stringify(input));
-
-        let prediction;
-        try {
-          const { stdout, stderr } = await execFileAsync(
-            getPythonExecutable(),
-            [analyzePyPath, "predict_file", tempFile],
-            {
-              timeout: 30000
-            }
-          );
-          prediction = JSON.parse(stdout.trim());
-          if (prediction.error) {
-            return res.status(400).json({
-              message: prediction.error
-            });
-          }
-        } catch (error: any) {
-          if (error.killed || error.signal === "SIGTERM") {
-            return res.status(408).json({
-              message: "Clinical assessment preview timed out."
-            });
-          }
-          console.warn("Python prediction preview failed, running clinical rule-based fallback:", error);
-          prediction = calculateClinicalFallback(input);
-        }
-        console.log(`[AUDIT] preview requested by=${req.session.user?.email} riskCategory=${prediction.riskCategory} riskScore=${prediction.riskScore} at=${new Date().toISOString()}`);
-        return res.json({
-          riskScore: prediction.riskScore,
-          riskCategory: prediction.riskCategory,
-          factors: prediction.factors ?? [],
-          confidenceInterval: prediction.confidenceInterval ?? null,
-          modelConfidence: prediction.modelConfidence ?? null
-        });
-      } catch (err) {
-        if (err instanceof z.ZodError) {
-          return res.status(400).json({
-            message: err.errors[0].message
-          });
-        }
-        logger.error({ err }, "Error creating assessment preview");
-        return res.status(500).json({ message: "Internal server error" });
-      } finally {
-        try {
-          await unlink(tempFilePath);
-        } catch (e) {
-          console.warn("Failed to clean up temp file (preview):", tempFilePath, e);
-        }
-      }
-    }
-  );
-
-  app.post(
-    api.assessments.create.path,
-    requireAuth,
-    requireVerified,
-    assessmentLimiter,
-    async (req, res) => {
-      const userId = req.session.user?.email;
-      if (!userId) {
-        return res.status(401).json({
-          message: "Authentication required.",
-        });
-      }
-
-      let requestFingerprint: string | null = null;
-      let tempFile: string | null = null;
-
-      try {
-        const input = api.assessments.create.input.parse(req.body);
-        requestFingerprint = generateRequestFingerprint(input, userId);
-
-        if (activeInferenceRequests.has(requestFingerprint)) {
-          return res.status(409).json({
-            message: "An identical assessment request is already being processed."
-          });
-        }
-
-        activeInferenceRequests.add(requestFingerprint);
-
-        tempFile = path.join(
-          os.tmpdir(),
-          `${randomUUID()}.json`
-        );
-
-        await writeFile(tempFile, JSON.stringify(input));
-
-        let prediction;
-        let isFallback = false;
-
-        try {
-          const { stdout, stderr } = await execFileAsync(
-            getPythonExecutable(),
-            [analyzePyPath, "predict_file", tempFile],
-            {
-              timeout: 30000
-            }
-          );
-          prediction = JSON.parse(stdout.trim());
-          if (prediction.error) {
-            return res.status(400).json({
-              message: prediction.error
-            });
-          }
-        } catch (error: any) {
-          if (error.killed || error.signal === "SIGTERM") {
-            return res.status(408).json({
-              message: "Clinical assessment generation timed out."
-            });
-          }
-          console.warn("Python ML prediction failed, running clinical rule-based fallback:", error);
-          prediction = calculateClinicalFallback(input);
-          isFallback = true;
-        }
-
-        prediction.disclaimer =
-          "DISCLAIMER: This is a clinical decision support tool and is not a medical diagnosis. Please consult with a healthcare professional for clinical decisions." +
-          (isFallback
-            ? " (Generated via fallback rule-based clinical support model due to system unavailability)"
-            : "");
-        prediction.isFallback = isFallback;
-
-        const assessment = await storage.createAssessment({
-          ...input,
-          riskScore: Number(prediction.riskScore),
-          riskCategory: prediction.riskCategory,
-          factors: prediction.factors,
-          confidenceInterval: prediction.confidenceInterval ?? null,
-          modelConfidence:
-            prediction.modelConfidence == null
-              ? undefined
-              : Number(prediction.modelConfidence),
-          createdBy: userId
-        });
-        console.log(`[AUDIT] prediction created by=${userId} riskCategory=${prediction.riskCategory} riskScore=${prediction.riskScore} at=${new Date().toISOString()}`);
-        return res.status(201).json({
-          ...assessment,
-          prediction
-        });
-
-        // Trigger automated email alert if risk score is critical (> 80%)
-        if (assessment.riskScore > 80) {
-          try {
-            const user = await storage.getUserById(userId);
-            if (user && user.email) {
-              await sendCriticalRiskAlert(
-                user.email,
-                assessment.patientName || "Unknown Patient",
-                assessment.riskScore,
-                assessment.id
-              );
-            }
-          } catch (emailErr) {
-            console.error("Failed to send critical risk email alert:", emailErr);
-          }
-        }
-
-        return res.status(201).json({ ...assessment, prediction });
-      } catch (err) {
-        if (err instanceof z.ZodError) {
-          return res.status(400).json({
-            message: err.errors[0].message
-          });
-        }
-        logger.error({ err }, "Error creating assessment");
-        return res
-          .status(500)
-          .json({ message: "Failed to generate clinical assessment." });
-      } finally {
-        if (tempFile) {
-          try {
-            await unlink(tempFilePath);
-          } catch (e) {
-            console.warn("Failed to clean up temp file (create):", tempFilePath, e);
-          }
-        }
-        if (requestFingerprint) {
-          activeInferenceRequests.delete(requestFingerprint);
-        }
-      }
-    }
-  );
-
-  app.post(
-    "/api/assessments/bulk",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      const userId = (req.session.user as any)?.id;
-      if (!userId) {
-        return res.status(401).json({ message: "Authentication required." });
-      }
-
-      const inputSchema = z.array(api.assessments.create.input);
-      let tempFilePath: string | null = null;
-      let requestFingerprint: string | null = null;
-
-      try {
-        const input = inputSchema.parse(req.body.assessments);
-        
-        requestFingerprint = generateRequestFingerprint(input, userId);
-        if (activeInferenceRequests.has(requestFingerprint)) {
-          return res.status(409).json({ message: "Bulk request already processing." });
-        }
-        activeInferenceRequests.add(requestFingerprint);
-
-        tempFilePath = path.join(os.tmpdir(), `bulk_${randomUUID()}.json`);
-        await writeFile(tempFilePath, JSON.stringify(input));
-
-        let predictions: any[];
-        try {
-          const { stdout } = await execFileAsync(
-            getPythonExecutable(),
-            [analyzePyPath, "predict_file", tempFilePath],
-            { timeout: 60000, maxBuffer: 50 * 1024 * 1024 }
-          );
-
-          predictions = JSON.parse(stdout.trim());
-          if (!Array.isArray(predictions)) {
-            throw new Error("Expected array of predictions");
-          }
-        } catch (error: any) {
-          return res.status(500).json({ message: "Bulk ML processing failed or timed out." });
-        }
-
-        const createdAssessments = await Promise.all(
-          input.map((assessment, index) => {
-            const prediction = predictions[index];
-            return storage.createAssessment({
-              ...assessment,
-              riskScore: Number(prediction.riskScore),
-              riskCategory: prediction.riskCategory,
-              factors: prediction.factors,
-              confidenceInterval: prediction.confidenceInterval ?? null,
-              modelConfidence: prediction.modelConfidence == null ? undefined : Number(prediction.modelConfidence),
-              createdBy: userId,
-            });
-          })
-        );
-
-        return res.status(201).json({ count: createdAssessments.length, assessments: createdAssessments });
-      } catch (err) {
-        if (err instanceof z.ZodError) {
-          return res.status(400).json({ message: "Invalid bulk input data format. Ensure all rows meet schema requirements." });
-        }
-        console.error("Bulk create error:", err);
-        return res.status(500).json({ message: "Failed to generate bulk assessments." });
-      } finally {
-        if (tempFilePath) {
-          try { await unlink(tempFilePath); } catch {}
-        }
-        if (requestFingerprint) {
-          activeInferenceRequests.delete(requestFingerprint);
-        }
-      }
-    }
-  );
-
-  app.post(
-    "/api/assessments/bulk",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      const userId = (req.session.user as any)?.id;
-      if (!userId) {
-        return res.status(401).json({ message: "Authentication required." });
-      }
-
-      const inputSchema = z.array(api.assessments.create.input);
-      let tempFilePath: string | null = null;
-      let requestFingerprint: string | null = null;
-
-      try {
-        const input = inputSchema.parse(req.body.assessments);
-        
-        requestFingerprint = generateRequestFingerprint(input, userId);
-        if (activeInferenceRequests.has(requestFingerprint)) {
-          return res.status(409).json({ message: "Bulk request already processing." });
-        }
-        activeInferenceRequests.add(requestFingerprint);
-
-        tempFilePath = path.join(os.tmpdir(), `bulk_${randomUUID()}.json`);
-        await writeFile(tempFilePath, JSON.stringify(input));
-
-        let predictions: any[];
-        try {
-          const { stdout } = await execFileAsync(
-            getPythonExecutable(),
-            [analyzePyPath, "predict_file", tempFilePath],
-            { timeout: 60000, maxBuffer: 50 * 1024 * 1024 }
-          );
-
-          predictions = JSON.parse(stdout.trim());
-          if (!Array.isArray(predictions)) {
-            throw new Error("Expected array of predictions");
-          }
-        } catch (error: any) {
-          return res.status(500).json({ message: "Bulk ML processing failed or timed out." });
-        }
-
-        const createdAssessments = await Promise.all(
-          input.map((assessment, index) => {
-            const prediction = predictions[index];
-            return storage.createAssessment({
-              ...assessment,
-              riskScore: Number(prediction.riskScore),
-              riskCategory: prediction.riskCategory,
-              factors: prediction.factors,
-              confidenceInterval: prediction.confidenceInterval ?? null,
-              modelConfidence: prediction.modelConfidence == null ? undefined : Number(prediction.modelConfidence),
-              createdBy: userId,
-            });
-          })
-        );
-
-        return res.status(201).json({ count: createdAssessments.length, assessments: createdAssessments });
-      } catch (err) {
-        if (err instanceof z.ZodError) {
-          return res.status(400).json({ message: "Invalid bulk input data format. Ensure all rows meet schema requirements." });
-        }
-        console.error("Bulk create error:", err);
-        return res.status(500).json({ message: "Failed to generate bulk assessments." });
-      } finally {
-        if (tempFilePath) {
-          try { await unlink(tempFilePath); } catch {}
-        }
-        if (requestFingerprint) {
-          activeInferenceRequests.delete(requestFingerprint);
-        }
-      }
-    }
-  );
-
-  app.get(api.assessments.list.path, requireAuth, requireVerified, async (req, res) => {
-    try {
-      const userEmail = req.session.user?.email;
-      const cursorStr = req.query.cursor as string;
-      const limitStr = req.query.limit as string;
-      const cursor = cursorStr ? parseInt(cursorStr, 10) : undefined;
-      const limit = limitStr ? parseInt(limitStr, 10) : 50;
-
-      const assessments = await storage.getAssessments(limit, cursor, userEmail);
-
-      const nextCursor = assessments.length === limit && assessments.length > 0 
-        ? assessments[assessments.length - 1].id 
-        : undefined;
-
-      res.json({ data: assessments, nextCursor });
-
-    } catch (err) {
-      res.status(500).json({
-        message: "Failed to fetch assessments"
-      });
-    }
-  });
-  app.get(
-    "/api/assessments/export.csv",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const userEmail = req.session.user?.email;
-        const assessments = await storage.getAssessments(1000, undefined, userEmail);
-
-        const csv = assessmentsToCsv(
-          assessments as unknown as Record<string, unknown>[]
-        );
-
-        res.header("Content-Type", "text/csv");
-        res.attachment("assessments.csv");
-        return res.send(csv);
-      } catch (err) {
-        console.error("Export error:", err);
-        return res.status(500).json({ message: "Failed to export data" });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/search
-   *
-   * Secure patient/assessment search endpoint.
-   *
-   * Security controls:
-   * 1. PRIMARY: Drizzle ORM ilike()/eq() — query parameters are bound placeholders,
-   *    never interpolated into raw SQL strings.  This prevents SQL injection.
-   * 2. SUPPLEMENTARY: Zod schema validates input length, character set, and rejects
-   *    known injection signatures before the query is even constructed.
-   * 3. Security logging: suspicious patterns are logged (without PHI) for audit.
-   * 4. User scoping: results are always filtered to the authenticated user's records.
-   * 5. Generic errors: DB errors are sanitized — no table names or SQL syntax leaked.
-   *
-   * Query params:
-   *   q            - search term (max 200 chars, safe characters only)
-   *   riskCategory - optional: LOW | MODERATE | HIGH
-   *   page         - page number (default 1)
-   *   limit        - results per page, max 100 (default 20)
-   */
-  app.get(
-    "/api/assessments/search",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const parseResult = searchQuerySchema.safeParse(req.query);
-
-        if (!parseResult.success) {
-          const rawQ = typeof req.query.q === "string" ? req.query.q : "";
-          const analysis = analyzeSearchInput(rawQ);
-
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SQL_INJECTION_ATTEMPT",
-              "Injection-like pattern detected in search query parameter",
-              req,
-              {
-                matchedPattern: analysis.pattern,
-                userId: req.session.user?.id,
-              }
-            );
-          } else {
-            logSecurityEvent(
-              "MALFORMED_SEARCH_QUERY",
-              "Search query failed validation",
-              req,
-              { userId: req.session.user?.id }
-            );
-          }
-
-          return res.status(400).json({
-            message: parseResult.error.errors[0]?.message ?? "Invalid search parameters.",
-          });
-        }
-
-        const { q, riskCategory, cursor, limit } = parseResult.data;
-        const userEmail = req.session.user?.email;
-
-        if (q) {
-          const analysis = analyzeSearchInput(q);
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SUSPICIOUS_SEARCH_PATTERN",
-              "Validated search term contains a suspicious pattern",
-              req,
-              { matchedPattern: analysis.pattern, userId: req.session.user?.id }
-            );
-          }
-        }
-
-        const results = await storage.searchAssessments(
-          q ?? "",
-          userEmail,
-          riskCategory,
-          limit,
-          cursor
-        );
-
-        const nextCursor = results.length === limit && results.length > 0 
-          ? results[results.length - 1].id 
-          : undefined;
-
-        return res.json({ data: results, nextCursor });
-
-      } catch (err) {
-        console.error("Assessment search error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  app.get(
-    "/api/assessments/analytics",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const userEmail = req.session.user?.email;
-        if (!userEmail) {
-           return res.status(401).json({ message: "Unauthorized" });
-        }
-        const stats = await storage.getAnalyticsStats(userEmail);
-        return res.json(stats);
-      } catch (err) {
-        console.error("Analytics fetch error:", err);
-        return res.status(500).json({ message: "Failed to fetch analytics" });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Object-level authorization is enforced explicitly before returning records.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Object-Level Authorization Check
-        if (!canAccessPatientRecord(user, assessment)) {
-          // Log unauthorized access attempt (IDOR/Enumeration attempt)
-          logAccessAttempt(
-            user.id,
-            "Assessment",
-            id,
-            false,
-            "IDOR attempt: User not authorized to access this patient record"
-          );
-          
-          // Return 404 to prevent ID enumeration
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Authorized access
-        logAccessAttempt(user.id, "Assessment", id, true, "Authorized access");
-        return res.json(assessment);
-
-      } catch (err) {
-        // 4. Sanitize DB errors — never expose table names, SQL syntax, or stack traces
-        logger.error({ err }, "Assessment search error");
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Results are scoped to the authenticated user to prevent cross-user data access.
-   *
-   * Security: uses Drizzle ORM eq() with bound parameters — not string-concatenated.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        if (!canAccessPatientRecord(user, assessment)) {
-          logAccessAttempt(user.id, "Assessment", id, false, "IDOR attempt: User not authorized to access this patient record");
-
-          // Return 404 to prevent ID enumeration
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Authorized access
-        logAccessAttempt((user as any).id, "Assessment", id, true, "Authorized access");
-        return res.json(assessment);
-
-      } catch (err) {
-        console.error("Assessment fetch error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/search
-   *
-   * Secure patient/assessment search endpoint.
-   *
-   * Security controls:
-   * 1. PRIMARY: Drizzle ORM ilike()/eq() — query parameters are bound placeholders,
-   *    never interpolated into raw SQL strings.  This prevents SQL injection.
-   * 2. SUPPLEMENTARY: Zod schema validates input length, character set, and rejects
-   *    known injection signatures before the query is even constructed.
-   * 3. Security logging: suspicious patterns are logged (without PHI) for audit.
-   * 4. User scoping: results are always filtered to the authenticated user's records.
-   * 5. Generic errors: DB errors are sanitized — no table names or SQL syntax leaked.
-   *
-   * Query params:
-   *   q            - search term (max 200 chars, safe characters only)
-   *   riskCategory - optional: LOW | MODERATE | HIGH
-   *   page         - page number (default 1)
-   *   limit        - results per page, max 100 (default 20)
-   */
-  app.get(
-    "/api/assessments/search",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        // 1. Validate and parse query parameters
-        const parseResult = searchQuerySchema.safeParse(req.query);
-
-        if (!parseResult.success) {
-          // Check whether the failure looks like an injection attempt
-          const rawQ = typeof req.query.q === "string" ? req.query.q : "";
-          const analysis = analyzeSearchInput(rawQ);
-
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SQL_INJECTION_ATTEMPT",
-              "Injection-like pattern detected in search query parameter",
-              req,
-              {
-                matchedPattern: analysis.pattern,
-                userId: req.session.user?.id,
-              }
-            );
-          } else {
-            logSecurityEvent(
-              "MALFORMED_SEARCH_QUERY",
-              "Search query failed validation",
-              req,
-              { userId: req.session.user?.id }
-            );
-          }
-
-          return res.status(400).json({
-            message: parseResult.error.errors[0]?.message ?? "Invalid search parameters.",
-          });
-        }
-
-        const { q, riskCategory, page, limit } = parseResult.data;
-        const offset = (page - 1) * limit;
-        const userEmail = req.session.user?.email;
-
-        // 2. Log suspicious-but-valid patterns for monitoring
-        if (q) {
-          const analysis = analyzeSearchInput(q);
-          if (!analysis.safe) {
-            // Validation already rejected this above, but log defensively
-            logSecurityEvent(
-              "SUSPICIOUS_SEARCH_PATTERN",
-              "Validated search term contains a suspicious pattern",
-              req,
-              { matchedPattern: analysis.pattern, userId: req.session.user?.id }
-            );
-          }
-        }
-
-        // 3. Execute parameterized search — Drizzle ORM sends $1, $2 … placeholders
-        const results = await storage.searchAssessments(
-          q ?? "",
-          userEmail,
-          riskCategory,
-          limit,
-          offset
-        );
-        logAccessAttempt(req.session.user!.id, "Assessments", "search", true, "Searched assessments");
-
-        return res.json(results);
-
-      } catch (err) {
-        // 4. Sanitize DB errors — never expose table names, SQL syntax, or stack traces
-        console.error("Assessment search error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Results are scoped to the authenticated user to prevent cross-user data access.
-   *
-   * Security: uses Drizzle ORM eq() with bound parameters — not string-concatenated.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        if (!canAccessPatientRecord(user, assessment)) {
-          logAccessAttempt(user.id, "Assessment", id, false, "IDOR attempt: User not authorized to access this patient record");
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        return res.json(assessment);
-
-      } catch (err) {
-        console.error("Assessment fetch error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/search
-   *
-   * Secure patient/assessment search endpoint.
-   *
-   * Security controls:
-   * 1. PRIMARY: Drizzle ORM ilike()/eq() — query parameters are bound placeholders,
-   *    never interpolated into raw SQL strings.  This prevents SQL injection.
-   * 2. SUPPLEMENTARY: Zod schema validates input length, character set, and rejects
-   *    known injection signatures before the query is even constructed.
-   * 3. Security logging: suspicious patterns are logged (without PHI) for audit.
-   * 4. User scoping: results are always filtered to the authenticated user's records.
-   * 5. Generic errors: DB errors are sanitized — no table names or SQL syntax leaked.
-   *
-   * Query params:
-   *   q            - search term (max 200 chars, safe characters only)
-   *   riskCategory - optional: LOW | MODERATE | HIGH
-   *   page         - page number (default 1)
-   *   limit        - results per page, max 100 (default 20)
-   */
-  app.get(
-    "/api/assessments/search",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        // 1. Validate and parse query parameters
-        const parseResult = searchQuerySchema.safeParse(req.query);
-
-        if (!parseResult.success) {
-          // Check whether the failure looks like an injection attempt
-          const rawQ = typeof req.query.q === "string" ? req.query.q : "";
-          const analysis = analyzeSearchInput(rawQ);
-
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SQL_INJECTION_ATTEMPT",
-              "Injection-like pattern detected in search query parameter",
-              req,
-              {
-                matchedPattern: analysis.pattern,
-                userId: req.session.user?.id,
-              }
-            );
-          } else {
-            logSecurityEvent(
-              "MALFORMED_SEARCH_QUERY",
-              "Search query failed validation",
-              req,
-              { userId: req.session.user?.id }
-            );
-          }
-
-          return res.status(400).json({
-            message: parseResult.error.errors[0]?.message ?? "Invalid search parameters.",
-          });
-        }
-
-        const { q, riskCategory, page, limit } = parseResult.data;
-        const offset = (page - 1) * limit;
-        const userEmail = req.session.user?.email;
-
-        // 2. Log suspicious-but-valid patterns for monitoring
-        if (q) {
-          const analysis = analyzeSearchInput(q);
-          if (!analysis.safe) {
-            // Validation already rejected this above, but log defensively
-            logSecurityEvent(
-              "SUSPICIOUS_SEARCH_PATTERN",
-              "Validated search term contains a suspicious pattern",
-              req,
-              { matchedPattern: analysis.pattern, userId: req.session.user?.id }
-            );
-          }
-        }
-
-        // 3. Execute parameterized search — Drizzle ORM sends $1, $2 … placeholders
-        const results = await storage.searchAssessments(
-          q ?? "",
-          userEmail,
-          riskCategory,
-          limit,
-          offset
-        );
-        logAccessAttempt(req.session.user!.id, "Assessments", "search", true, "Searched assessments");
-
-        return res.json(results);
-
-      } catch (err) {
-        // 4. Sanitize DB errors — never expose table names, SQL syntax, or stack traces
-        console.error("Assessment search error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Results are scoped to the authenticated user to prevent cross-user data access.
-   *
-   * Security: uses Drizzle ORM eq() with bound parameters — not string-concatenated.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        // Fetch the record regardless of owner
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          // Normal 404
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Object-Level Authorization Check
-        if (!canAccessPatientRecord(user, assessment)) {
-          // Log unauthorized access attempt (IDOR/Enumeration attempt)
-          logAccessAttempt(
-            user.id,
-            "Assessment",
-            id,
-            false,
-            "IDOR attempt: User not authorized to access this patient record"
-          );
-          
-          // Return 404 to prevent ID enumeration
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Authorized access
-        logAccessAttempt(user.id, "Assessment", id, true, "Authorized access");
-        return res.json(assessment);
-
-      } catch (err) {
-        console.error("Assessment fetch error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/search
-   *
-   * Secure patient/assessment search endpoint.
-   *
-   * Security controls:
-   * 1. PRIMARY: Drizzle ORM ilike()/eq() — query parameters are bound placeholders,
-   *    never interpolated into raw SQL strings.  This prevents SQL injection.
-   * 2. SUPPLEMENTARY: Zod schema validates input length, character set, and rejects
-   *    known injection signatures before the query is even constructed.
-   * 3. Security logging: suspicious patterns are logged (without PHI) for audit.
-   * 4. User scoping: results are always filtered to the authenticated user's records.
-   * 5. Generic errors: DB errors are sanitized — no table names or SQL syntax leaked.
-   *
-   * Query params:
-   *   q            - search term (max 200 chars, safe characters only)
-   *   riskCategory - optional: LOW | MODERATE | HIGH
-   *   page         - page number (default 1)
-   *   limit        - results per page, max 100 (default 20)
-   */
-  app.get(
-    "/api/assessments/search",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        // 1. Validate and parse query parameters
-        const parseResult = searchQuerySchema.safeParse(req.query);
-
-        if (!parseResult.success) {
-          // Check whether the failure looks like an injection attempt
-          const rawQ = typeof req.query.q === "string" ? req.query.q : "";
-          const analysis = analyzeSearchInput(rawQ);
-
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SQL_INJECTION_ATTEMPT",
-              "Injection-like pattern detected in search query parameter",
-              req,
-              {
-                matchedPattern: analysis.pattern,
-                userId: req.session.user?.id,
-              }
-            );
-          } else {
-            logSecurityEvent(
-              "MALFORMED_SEARCH_QUERY",
-              "Search query failed validation",
-              req,
-              { userId: req.session.user?.id }
-            );
-          }
-
-          return res.status(400).json({
-            message: parseResult.error.errors[0]?.message ?? "Invalid search parameters.",
-          });
-        }
-
-        const { q, riskCategory, page, limit } = parseResult.data;
-        const offset = (page - 1) * limit;
-        const userEmail = req.session.user?.email;
-
-        // 2. Log suspicious-but-valid patterns for monitoring
-        if (q) {
-          const analysis = analyzeSearchInput(q);
-          if (!analysis.safe) {
-            // Validation already rejected this above, but log defensively
-            logSecurityEvent(
-              "SUSPICIOUS_SEARCH_PATTERN",
-              "Validated search term contains a suspicious pattern",
-              req,
-              { matchedPattern: analysis.pattern, userId: req.session.user?.id }
-            );
-          }
-        }
-
-        // 3. Execute parameterized search — Drizzle ORM sends $1, $2 … placeholders
-        const results = await storage.searchAssessments(
-          q ?? "",
-          userEmail,
-          riskCategory,
-          limit,
-          offset
-        );
-        logAccessAttempt(req.session.user!.id, "Assessments", "search", true, "Searched assessments");
-
-        return res.json(results);
-
-      } catch (err) {
-        // 4. Sanitize DB errors — never expose table names, SQL syntax, or stack traces
-        console.error("Assessment search error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Results are scoped to the authenticated user to prevent cross-user data access.
-   *
-   * Security: uses Drizzle ORM eq() with bound parameters — not string-concatenated.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        // Fetch the record regardless of owner
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          // Normal 404
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Object-Level Authorization Check
-        if (!canAccessPatientRecord(user, assessment)) {
-          // Log unauthorized access attempt (IDOR/Enumeration attempt)
-          logAccessAttempt(
-            user.id,
-            "Assessment",
-            id,
-            false,
-            "IDOR attempt: User not authorized to access this patient record"
-          );
-          
-          // Return 404 to prevent ID enumeration
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Authorized access
-        logAccessAttempt(user.id, "Assessment", id, true, "Authorized access");
-        return res.json(assessment);
-
-      } catch (err) {
-        console.error("Assessment fetch error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/search
-   *
-   * Secure patient/assessment search endpoint.
-   *
-   * Security controls:
-   * 1. PRIMARY: Drizzle ORM ilike()/eq() — query parameters are bound placeholders,
-   *    never interpolated into raw SQL strings.  This prevents SQL injection.
-   * 2. SUPPLEMENTARY: Zod schema validates input length, character set, and rejects
-   *    known injection signatures before the query is even constructed.
-   * 3. Security logging: suspicious patterns are logged (without PHI) for audit.
-   * 4. User scoping: results are always filtered to the authenticated user's records.
-   * 5. Generic errors: DB errors are sanitized — no table names or SQL syntax leaked.
-   *
-   * Query params:
-   *   q            - search term (max 200 chars, safe characters only)
-   *   riskCategory - optional: LOW | MODERATE | HIGH
-   *   page         - page number (default 1)
-   *   limit        - results per page, max 100 (default 20)
-   */
-  app.get(
-    "/api/assessments/search",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        // 1. Validate and parse query parameters
-        const parseResult = searchQuerySchema.safeParse(req.query);
-
-        if (!parseResult.success) {
-          // Check whether the failure looks like an injection attempt
-          const rawQ = typeof req.query.q === "string" ? req.query.q : "";
-          const analysis = analyzeSearchInput(rawQ);
-
-          if (!analysis.safe) {
-            logSecurityEvent(
-              "SQL_INJECTION_ATTEMPT",
-              "Injection-like pattern detected in search query parameter",
-              req,
-              {
-                matchedPattern: analysis.pattern,
-                userId: req.session.user?.id,
-              }
-            );
-          } else {
-            logSecurityEvent(
-              "MALFORMED_SEARCH_QUERY",
-              "Search query failed validation",
-              req,
-              { userId: req.session.user?.id }
-            );
-          }
-
-          return res.status(400).json({
-            message: parseResult.error.errors[0]?.message ?? "Invalid search parameters.",
-          });
-        }
-
-        const { q, riskCategory, page, limit } = parseResult.data;
-        const offset = (page - 1) * limit;
-        const userEmail = req.session.user?.email;
-
-        // 2. Log suspicious-but-valid patterns for monitoring
-        if (q) {
-          const analysis = analyzeSearchInput(q);
-          if (!analysis.safe) {
-            // Validation already rejected this above, but log defensively
-            logSecurityEvent(
-              "SUSPICIOUS_SEARCH_PATTERN",
-              "Validated search term contains a suspicious pattern",
-              req,
-              { matchedPattern: analysis.pattern, userId: req.session.user?.id }
-            );
-          }
-        }
-
-        // 3. Execute parameterized search — Drizzle ORM sends $1, $2 … placeholders
-        const results = await storage.searchAssessments(
-          q ?? "",
-          userEmail,
-          riskCategory,
-          limit,
-          offset
-        );
-        logAccessAttempt(req.session.user!.id, "Assessments", "search", true, "Searched assessments");
-
-        return res.json(results);
-
-      } catch (err) {
-        // 4. Sanitize DB errors — never expose table names, SQL syntax, or stack traces
-        console.error("Assessment search error:", err);
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  /**
-   * GET /api/assessments/:id
-   *
-   * Fetch a single assessment by numeric ID.
-   * Results are scoped to the authenticated user to prevent cross-user data access.
-   *
-   * Security: uses Drizzle ORM eq() with bound parameters — not string-concatenated.
-   */
-  app.get(
-    "/api/assessments/:id",
-    requireAuth,
-    requireVerified,
-    async (req, res) => {
-      try {
-        const id = parseInt(req.params.id, 10);
-
-        if (isNaN(id) || id <= 0) {
-          return res.status(400).json({ message: "Invalid assessment ID." });
-        }
-
-        const user = req.session.user;
-        if (!user) {
-          return res.status(401).json({ message: "Unauthorized" });
-        }
-
-        // Fetch the record regardless of owner
-        const assessment = await storage.getAssessmentById(id);
-
-        if (!assessment) {
-          // Normal 404
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Object-Level Authorization Check
-        if (!canAccessPatientRecord(user, assessment)) {
-          // Log unauthorized access attempt (IDOR/Enumeration attempt)
-          logAccessAttempt(
-            user.id,
-            "Assessment",
-            id,
-            false,
-            "IDOR attempt: User not authorized to access this patient record"
-          );
-          
-          // Return 404 to prevent ID enumeration
-          return res.status(404).json({ message: "Assessment not found." });
-        }
-
-        // Authorized access
-        logAccessAttempt(user.id, "Assessment", id, true, "Authorized access");
-        return res.json(assessment);
-
-      } catch (err) {
-        logger.error({ err }, "Assessment fetch error");
-        const { statusCode, message } = sanitizeDatabaseError(err);
-        return res.status(statusCode).json({ message });
-      }
-    }
-  );
-
-  // ─── Admin Routes ────────────────────────────────────────────────
-
+  app.use("/api/assessments", mlRouter);
+  app.use("/api/assessments", exportsRouter);
+  app.use("/api/assessments", analyticsRouter);
   app.get("/api/admin/users", requireAuth, requireAdmin, async (req, res) => {
     try {
       const page = Math.max(1, parseInt(req.query.page as string) || 1);
